refactor: extract openai wire model
